info_retrieval_agent/requirements.txt
# requirements.txt for info_retrieval_agent
fastapi
uvicorn
python-dotenv
pydantic
# Add other necessary libraries like pinecone-client
